quart/requirements.txt
asyncpg
quart
quart-cors
pytest-asyncio
pytest-dotenv
uvicorn